chore: longitude, latitude의 data type을 double로 변경

+++ gusto/src/main/java/com/umc/gusto/domain/store/model/response/GetStoresInMapResponse.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class GetStoresInMapResponse{
     Long storeId;
     String storeName;
-    Float longitude;
-    Float latitude;
+    Double longitude;
+    Double latitude;
 }
